refactor: rename `Lists` to `ArtworkLists`
