+++ src/Schema/index.ts
@@ -1,5 +1,6 @@
 export * from "./Events"
 export * from "./Events/AddToCalendar"
+export * from "./Events/ArtworkLists"
 export * from "./Events/Authentication"
 export * from "./Events/Click"
 export * from "./Events/CollectorProfile"
@@ -9,7 +10,6 @@ export * from "./Events/FilterAndSort"
 export * from "./Events/Impression"
 export * from "./Events/ImpressionTracking"
 export * from "./Events/Inquiries"
-export * from "./Events/Lists"
 export * from "./Events/MyCollection"
 export * from "./Events/MyCollectionInsights"
 export * from "./Events/ReverseImageSearch"
